chore(core): move summary title v2
