+++ mobile-app/app/components/SummaryTitle.tsx
@@ -1,62 +1,156 @@
 import BigNumber from "bignumber.js";
-
 import { NumericFormat as NumberFormat } from "react-number-format";
 import { tailwind } from "@tailwind";
-import { ViewProps } from "react-native";
-import { ThemedText } from "./themed";
+import { getNativeIcon } from "@components/icons/assets";
+import { translate } from "@translations";
+import { RandomAvatar } from "@screens/AppNavigator/screens/Portfolio/components/RandomAvatar";
+import { AddressType } from "@store/wallet";
 import { View } from ".";
-
-type SummaryTitleProps = React.PropsWithChildren<ViewProps> &
-  ISummaryTitleProps;
-type SuffixType = "text" | "component";
+import { ThemedTextV2, ThemedViewV2 } from "./themed";
 
 interface ISummaryTitleProps {
   title: string;
   amount: BigNumber;
-  suffixType: SuffixType;
-  suffix?: string;
   testID: string;
+  fromAddress?: string;
+  fromAddressLabel?: string | null;
+  toAddress?: string;
+  toAddressLabel?: string | null;
+  customToAddressTitle?: string;
+  iconA: string;
+  iconB?: string;
+  addressType?: AddressType;
+  amountTextStyle?: string;
 }
 
-export function SummaryTitle(props: SummaryTitleProps): JSX.Element {
+export function SummaryTitle(props: ISummaryTitleProps): JSX.Element {
+  const IconA = getNativeIcon(props.iconA);
+  const IconB =
+    props.iconB !== undefined ? getNativeIcon(props.iconB) : undefined;
+
   return (
     <>
-      <ThemedText
-        dark={tailwind("text-gray-400")}
-        light={tailwind("text-gray-500")}
-        style={tailwind("text-sm")}
+      <ThemedTextV2
+        dark={tailwind("text-mono-dark-v2-500")}
+        light={tailwind("text-mono-light-v2-500")}
+        style={tailwind("text-xs font-normal-v2")}
         testID="confirm_title"
       >
         {props.title}
-      </ThemedText>
+      </ThemedTextV2>
 
-      <View style={tailwind("flex-row items-center")}>
-        <NumberFormat
-          displayType="text"
-          renderText={(value) => (
-            <ThemedText
-              style={tailwind("text-2xl font-bold flex-wrap pr-1")}
-              testID={props.testID}
-            >
-              {value}
-            </ThemedText>
+      <View style={tailwind("flex-col")}>
+        <View style={tailwind("flex-row items-center mt-2")}>
+          {IconB !== undefined ? (
+            <View style={tailwind("flex-row")}>
+              <IconA height={32} width={32} style={tailwind("z-10")} />
+              <IconB height={32} width={32} style={tailwind("-ml-3")} />
+            </View>
+          ) : (
+            <IconA height={32} width={32} />
           )}
-          thousandSeparator
-          value={props.amount.toFixed(8)}
-        />
 
-        {props.suffixType === "text" && (
-          <ThemedText
-            light={tailwind("text-gray-500")}
-            dark={tailwind("text-gray-400")}
-            style={tailwind("text-sm")}
-            testID={`${props.testID}_suffix`}
-          >
-            {props.suffix}
-          </ThemedText>
+          <NumberFormat
+            displayType="text"
+            renderText={(value) => (
+              <ThemedTextV2
+                style={tailwind(
+                  `text-3xl font-semibold-v2 flex-wrap pr-1 pl-2 ${
+                    props.amountTextStyle ?? ""
+                  }`
+                )}
+                testID={props.testID}
+              >
+                {value}
+              </ThemedTextV2>
+            )}
+            thousandSeparator
+            value={props.amount.toFixed(8)}
+          />
+        </View>
+
+        {props.fromAddress !== undefined && (
+          <View style={tailwind("flex-row items-center mt-5")}>
+            <ThemedTextV2
+              style={tailwind("text-xs font-normal-v2")}
+              dark={tailwind("text-mono-dark-v2-500")}
+              light={tailwind("text-mono-light-v2-500")}
+            >
+              {translate("screens/common", "From")}
+            </ThemedTextV2>
+            <ThemedViewV2
+              dark={tailwind("bg-mono-dark-v2-200")}
+              light={tailwind("bg-mono-light-v2-200")}
+              style={tailwind(
+                "rounded-full pl-1 pr-2.5 py-1 flex flex-row items-center overflow-hidden ml-2"
+              )}
+            >
+              <RandomAvatar name={props.fromAddress} size={20} />
+              <ThemedTextV2
+                ellipsizeMode="middle"
+                numberOfLines={1}
+                style={[
+                  tailwind("text-sm font-normal-v2 ml-1"),
+                  {
+                    minWidth: 10,
+                    maxWidth: 108,
+                  },
+                ]}
+                testID="wallet_address"
+              >
+                {props.fromAddressLabel ?? props.fromAddress}
+              </ThemedTextV2>
+            </ThemedViewV2>
+          </View>
         )}
 
-        {props.suffixType === "component" && props.children}
+        {props.toAddress !== undefined && props.addressType !== undefined && (
+          <View
+            style={tailwind("flex-row items-center mt-4")}
+            testID="summary_to_view"
+          >
+            <ThemedTextV2
+              style={tailwind("text-xs font-normal-v2")}
+              dark={tailwind("text-mono-dark-v2-500")}
+              light={tailwind("text-mono-light-v2-500")}
+            >
+              {props.customToAddressTitle ?? translate("screens/common", "To")}
+            </ThemedTextV2>
+            <ThemedViewV2
+              dark={tailwind("bg-mono-dark-v2-200")}
+              light={tailwind("bg-mono-light-v2-200")}
+              style={tailwind(
+                "flex flex-row items-center overflow-hidden rounded-full pr-2.5 py-1 ml-2",
+                {
+                  "pl-1": props.addressType === AddressType.WalletAddress,
+                  "pl-2.5": props.addressType !== AddressType.WalletAddress,
+                }
+              )}
+            >
+              {props.addressType === AddressType.WalletAddress && (
+                <View style={tailwind("mr-1")}>
+                  <RandomAvatar name={props.toAddress} size={20} />
+                </View>
+              )}
+              <ThemedTextV2
+                ellipsizeMode="middle"
+                numberOfLines={1}
+                style={[
+                  tailwind("text-sm font-normal-v2"),
+                  {
+                    minWidth: 10,
+                    maxWidth: 108,
+                  },
+                ]}
+                testID="summary_to_value"
+              >
+                {props.toAddressLabel != null && props.toAddressLabel.length > 0
+                  ? props.toAddressLabel
+                  : props.toAddress}
+              </ThemedTextV2>
+            </ThemedViewV2>
+          </View>
+        )}
       </View>
     </>
   );
