+++ mobile-app/app/screens/AppNavigator/screens/Dex/DexConfirmAddLiquidity.tsx
@@ -22,7 +22,7 @@ import { tailwind } from "@tailwind";
 import { translate } from "@translations";
 import { onTransactionBroadcast } from "@api/transaction/transaction_commands";
 import { useAppDispatch } from "@hooks/useAppDispatch";
-import { SummaryTitleV2 } from "@components/SummaryTitleV2";
+import { SummaryTitle } from "@components/SummaryTitle";
 import { useWalletContext } from "@shared-contexts/WalletContext";
 import { useAddressLabel } from "@hooks/useAddressLabel";
 import {
@@ -123,7 +123,7 @@ export function ConfirmAddLiquidityScreen({ route }: Props): JSX.Element {
   return (
     <ThemedScrollViewV2 style={tailwind("py-8 px-5")}>
       <ThemedViewV2 style={tailwind("flex-col pb-4 mb-4")}>
-        <SummaryTitleV2
+        <SummaryTitle
           iconA={pair.tokenA.displaySymbol}
           iconB={pair.tokenB.displaySymbol}
           fromAddress={address}
