+++ mobile-app/app/screens/AppNavigator/screens/Loans/screens/ConfirmBorrowLoanTokenScreen.tsx
@@ -28,7 +28,7 @@ import { useWalletContext } from "@shared-contexts/WalletContext";
 import { useWhaleApiClient } from "@shared-contexts/WhaleContext";
 import { getPrecisedCurrencyValue } from "@screens/AppNavigator/screens/Auctions/helpers/precision-token-value";
 import { useAppDispatch } from "@hooks/useAppDispatch";
-import { SummaryTitleV2 } from "@components/SummaryTitleV2";
+import { SummaryTitle } from "@components/SummaryTitle";
 import { useAddressLabel } from "@hooks/useAddressLabel";
 import { NumberRowV2 } from "@components/NumberRowV2";
 import { TextRowV2 } from "@components/TextRowV2";
@@ -101,7 +101,7 @@ export function ConfirmBorrowLoanTokenScreen({
 
   return (
     <ThemedScrollViewV2 contentContainerStyle={tailwind("px-5 py-8")}>
-      <SummaryTitleV2
+      <SummaryTitle
         amount={new BigNumber(borrowAmount)}
         title={translate(
           "screens/ConfirmBorrowLoanTokenScreen",
