+++ mobile-app/app/screens/AppNavigator/screens/Loans/screens/ConfirmEditCollateralScreen.tsx
@@ -29,7 +29,7 @@ import { useWalletContext } from "@shared-contexts/WalletContext";
 import { useWhaleApiClient } from "@shared-contexts/WhaleContext";
 import { getPrecisedCurrencyValue } from "@screens/AppNavigator/screens/Auctions/helpers/precision-token-value";
 import { useAppDispatch } from "@hooks/useAppDispatch";
-import { SummaryTitleV2 } from "@components/SummaryTitleV2";
+import { SummaryTitle } from "@components/SummaryTitle";
 import { useAddressLabel } from "@hooks/useAddressLabel";
 import { NumberRowV2 } from "@components/NumberRowV2";
 import { AddressType } from "@store/wallet";
@@ -172,7 +172,7 @@ export function ConfirmEditCollateralScreen({
       testID="confirm_edit_collateral_screen"
     >
       <View style={tailwind("mb-6")}>
-        <SummaryTitleV2
+        <SummaryTitle
           title={translate(
             "screens/ConfirmEditCollateralScreen",
             isAdd ? "You are adding collateral" : "You are removing collateral"
